requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
twilio==9.10.4
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.3
